backend/core/views.py
from rest_framework import viewsets

from backend.core.models import Document
from backend.core.serializers import DocumentSerializer


class DocumentViewSet(viewsets.ModelViewSet):
    queryset = Document.objects.all()
    serializer_class = DocumentSerializer
