requirements.txt
Django==4.2.*
django-extensions==3.2.*
djangorestframework==3.14.*
dr-scaffold==2.1.*
python-decouple==3.8.*
